refactor: move plugin business off core private imports
