+++ agent/plugin_composition/__init__.py
@@ -16,7 +16,7 @@ from agent.control.models import TurnItem, TurnItemKind, TurnStatus
 from agent.control.scoped_turn import TurnAcceptedReceipt
 from agent.control.timer import TimerHandle, TimerStatus
 from agent.tool_context import ToolExecutionContext
-from agent.turn_effects import PostCommitEffect, TurnStorage
+from agent.plugin_contracts.turn_effects import PostCommitEffect, TurnStorage
 from agent.plugin_composition.dashboard import DashboardContext
 from agent.plugin_composition.commands import (
     COMMANDS,
