+++ plugins/markdown_memory/message_plugin.py
@@ -25,8 +25,7 @@ from agent.plugin_composition import (
 from agent.plugin_composition.bindings import BINDINGS, Bindings
 from agent.plugin_composition.messages import MESSAGE_CATALOG
 from agent.plugin_composition.models import BoundChatModel, ChatModels, ContextLengthError, LLMResponse, ModelError
-from agent.llm_json import load_json_object_loose
-from infra.persistence.json_store import atomic_write_text
+from agent.plugin_contracts.json_store import atomic_write_text
 from agent.plugin_composition.messages import MessageCatalog
 from agent.plugin_contracts import ContentPart, Input, Message, Output, ToolResult
 from ._boundaries import (
@@ -34,6 +33,7 @@ from ._boundaries import (
     CompactionReader, ContentFacts, ContextBuilder, StoredSummary, SummaryLookup,
     TURN_PROJECTION, TurnProjection,
 )
+from .llm_json import load_json_object_loose
 
 if TYPE_CHECKING:
     from agent.plugin_composition.messages import MessageReader
