+++ plugins/scheduler/store.py
@@ -12,8 +12,8 @@ from zoneinfo import ZoneInfo, ZoneInfoNotFoundError
 
 from pydantic import BaseModel, ConfigDict, Field, model_validator
 
-from core.common.timekit import parse_iso as _parse_iso
-from infra.persistence.json_store import atomic_save_json
+from agent.plugin_contracts.json_store import atomic_save_json
+from agent.plugin_contracts.timekit import parse_iso as _parse_iso
 from .schedule import ScheduledJob, SCHEDULE_MAX_ACTIVE_JOBS, is_cron_expr, next_cron_fire
 
 _T = TypeVar("_T")
